requirements.txt
fastapi
uvicorn[standard]
neo4j
python-dotenv
google-generativeai
tqdm
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
sentence-transformers
torch
faiss-cpu
numpy